Renamer mappen Mottaker til InstitusjonOgVerge
